adding new architecture
