+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/LayerVisualizers/VectorLayerVisualizer.cs
@@ -14,26 +14,52 @@
 	public class VectorLayerVisualizer : LayerVisualizerBase
 	{
 		VectorSubLayerProperties _layerProperties;
+		public VectorSubLayerProperties SubLayerProperties
+		{
+			get
+			{
+				return _layerProperties;
+			}
+			set
+			{
+				_layerProperties = value;
+			}
+		}
+
+		public ModifierStackBase DefaultModifierStack
+		{
+			get
+			{
+				return _defaultStack;
+			}
+			set
+			{
+				_defaultStack = value;
+			}
+		}
+
 		LayerPerformanceOptions _performanceOptions;
 		private Dictionary<UnityTile, List<int>> _activeCoroutines;
 		int _entityInCurrentCoroutine = 0;
 
-		ModifierStackBase _defaultStack;
+		protected ModifierStackBase _defaultStack;
 		private HashSet<ulong> _activeIds;
 		private Dictionary<UnityTile, List<ulong>> _idPool; //necessary to keep _activeIds list up to date when unloading tiles
 
 		private string _key;
+
+
 
 		public override string Key
 		{
 			get { return _layerProperties.coreOptions.layerName; }
 			set { _layerProperties.coreOptions.layerName = value; }
 		}
+
 		public void SetProperties(VectorSubLayerProperties properties, LayerPerformanceOptions performanceOptions)
 		{
 			List<MeshModifier> defaultMeshModifierStack = new List<MeshModifier>();
-			List<GameObjectModifier> defaultGOModifierStack = new List<GameObjectModifier>();
-
+		 	List<GameObjectModifier> defaultGOModifierStack = new List<GameObjectModifier>();
 			_layerProperties = properties;
 			_performanceOptions = performanceOptions;
 
@@ -55,13 +81,6 @@
 			switch (properties.coreOptions.geometryType)
 			{
 				case VectorPrimitiveType.Point:
-					if(typeof(PrefabItemOptions).IsAssignableFrom(_layerProperties.GetType())) //to check that the instance is of type PrefabItemOptions
-					{
-						var itemProperties = (PrefabItemOptions)_layerProperties;
-						var prefabModifier = ScriptableObject.CreateInstance<PrefabModifier>();
-						prefabModifier.SetProperties(itemProperties.spawnPrefabOptions);
-						defaultGOModifierStack.Add(prefabModifier);
-					}
 					break;
 				case VectorPrimitiveType.Custom:
 					// Let the user add anything that they want
@@ -162,7 +181,6 @@
 			//Add any additional modifiers that were added.
 			_defaultStack.MeshModifiers.AddRange(_layerProperties.MeshModifiers);
 			_defaultStack.GoModifiers.AddRange(_layerProperties.GoModifiers);
-
 		}
 
 		public override void Initialize()
@@ -264,7 +282,7 @@
 					}
 				}
 
-				if (_performanceOptions.isEnabled && _entityInCurrentCoroutine >= _performanceOptions.entityPerCoroutine)
+				if (_performanceOptions!=null && _performanceOptions.isEnabled && _entityInCurrentCoroutine >= _performanceOptions.entityPerCoroutine)
 				{
 					_entityInCurrentCoroutine = 0;
 					yield return null;
