+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Modifiers/ModifierStackBase.cs
@@ -10,8 +10,8 @@ namespace Mapbox.Unity.MeshGeneration.Modifiers
 {
 	public class ModifierStackBase : ScriptableObject
 	{
-		[NodeEditorElement("Mesh Modifiers")] public List<MeshModifier> MeshModifiers;
-		[NodeEditorElement("Game Object Modifiers")] public List<GameObjectModifier> GoModifiers;
+		[NodeEditorElement("Mesh Modifiers")] public List<MeshModifier> MeshModifiers = new List<MeshModifier>();
+		[NodeEditorElement("Game Object Modifiers")] public List<GameObjectModifier> GoModifiers = new List<GameObjectModifier>();
 
 		public virtual GameObject Execute(UnityTile tile, VectorFeatureUnity feature, MeshData meshData, GameObject parent = null, string type = "")
 		{
